+++ scripts/lib/utilities/actorUtils.js
@@ -1,5 +1,5 @@
 import {socket, sockets} from '../sockets.js';
-import {genericUtils, socketUtils} from '../../utils.js';
+import {effectUtils, genericUtils, socketUtils} from '../../utils.js';
 import {ActorMedkit} from '../../applications/medkit-actor.js';
 function getEffects(actor) {
     return Array.from(actor.allApplicableEffects());
@@ -198,6 +198,20 @@ function getEquippedShield(actor) {
 function getAllEquippedArmor(actor) {
     return actor.items.find(i => Object.keys(CONFIG.DND5E.armorTypes).includes(i.system.type?.value) && i.system.equipped);
 }
+async function hasConditionBy(sourceActor, targetActor, statusId) {
+    let condition = effectUtils.getEffectByStatusID(targetActor, statusId);
+    if (!condition) return false;
+    let validKeys = ['macro.CE', 'macro.CUB', 'macro.StatusEffect', 'StatusEffect'];
+    let hasCondition = await actorUtils.getEffects(targetActor).find(async effect => {
+        let originItem = await effectUtils.getOriginItem(effect);
+        if (!originItem) return;
+        if (originItem?.actor != sourceActor) return;
+        if (effect.statuses.has(statusId)) return true;
+        if (effect.flags['chris-premades']?.conditions?.includes(statusId)) return true;
+        if (effect.changes.find(i => validKeys.includes(i.key) && i.value.toLowerCase() === statusId)) return true;
+    });
+    return hasCondition ? true : false;
+}
 export let actorUtils = {
     getEffects,
     addFavorites,
@@ -226,5 +240,6 @@ export let actorUtils = {
     getEquivalentSpellSlotName,
     getEquippedArmor,
     getEquippedShield,
-    getAllEquippedArmor
+    getAllEquippedArmor,
+    hasConditionBy
 };