+++ scripts/macros/2024/classFeatures/barbarian/worldTree/batteringRoots.js
@@ -1,5 +1,5 @@
-import {constants, dialogUtils, effectUtils, genericUtils, itemUtils} from '../../../../../utils.js';
-import {pushMastery, toppleMastery} from '../../../../mechanics/masteries.js';
+import {constants, dialogUtils, itemUtils} from '../../../../../utils.js';
+import {pushMastery, toppleMastery} from '../../../mechanics/masteries.js';
 async function turnStart({trigger: {entity: item}}) {
     let weaponProperties = itemUtils.getConfig(item, 'weaponProperties');
     if (!weaponProperties.length) return;